requirements.txt
google-genai==1.70.0
PyMuPDF==1.27.2.2
python-docx==1.2.0
pillow==12.2.0
python-dotenv==1.2.2
streamlit==1.56.0
